sdk/basyx/adapter/: redesigned adapter structure to match its use
