--- sdk/basyx/adapter/xml/__init__.py
@@ -1,11 +1,0 @@
-"""
-.. _adapter.xml.__init__:
-
-This package contains functionality for serialization and deserialization of BaSyx Python SDK objects into/from XML.
-
-:ref:`xml_serialization <adapter.xml.xml_serialization>`: The module offers a function to write an
-:class:`ObjectStore <basyx.aas.model.provider.AbstractObjectStore>` to a given file.
-
-:ref:`xml_deserialization <adapter.xml.xml_deserialization>`: The module offers a function to create an
-:class:`ObjectStore <basyx.aas.model.provider.AbstractObjectStore>` from a given xml document.
-"""
